Allow support for rating periods rather than a fixed rating period.
Earlier it was meant that a rating period is one tournament but this
ended up being really wrong. A rating period of multiple tournaments
is the way to go due to multiple regions that players cannot
participate on. This makes it fair for everyone without decaying their
RD significantly for not participating in a tournament outside their
region.